refactor: split combat-tracker hooks into combat/ and combatant-config hooks
The combat tracker is a Combat-level application, so its hooks belong with
Combat, not Combatant.

- Move the tracker hooks (renderCombatTracker + the combatant-row context menu
  building from getContextOptions) to src/document/combat/combat-tracker-hooks.ts.
- Split the combatant config-sheet field injection (moveFactor / displayedMedium)
  into src/document/combatant/combatant-config-hooks.ts — a Combatant concern.
- sohl.ts registers both; the test moves to tests/document/combat/.

No behavior change.
